requirements.txt
python-dotenv
requests
groq
fastapi
jinja2
uvicorn
pydantic
python-multipart